monstera_dbt_model/requirements.txt
# Core dbt dependencies
dbt-postgres==1.8.0
dbt-core==1.10.0rc2
dbt-adapters==1.16.0
dbt-common==1.25.0

# Database connectivity
psycopg2-binary==2.9.10

# Data generation and manipulation
faker==19.6.2
pandas==2.3.0
numpy==2.3.0

# System utilities
psutil==7.0.0

# Required for dbt functionality
Jinja2==3.1.6
PyYAML==6.0.2
click==8.2.1
agate==1.7.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
MarkupSafe==3.0.2
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==1.26.20

# dbt specific dependencies
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.8.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
msgpack==1.1.0
networkx==3.5
protobuf==5.29.5
snowplow-tracker==1.1.0
pytimeparse==1.1.8

# Additional utilities
more-itertools==10.7.0
parsedatetime==2.6
pathspec==0.11.2
python-slugify==8.0.4
text-unidecode==1.3
tzdata==2025.2
